Split EVM data types and factory into separate files
