+++ packages/utils/src/abi_encoder/calldata.ts
@@ -1,4 +1,3 @@
-
 import * as ethUtil from 'ethereumjs-util';
 import * as _ from 'lodash';
