+++ packages/utils/src/abi_encoder/index.ts
@@ -1,2 +1,2 @@
 export { EncodingRules, DecodingRules } from './calldata';
-export * from './evm_data_types';
+export * from './evm_data_type_factory';
